Rework platform detection into script; introduce port_std.
detect-platform facilitates identification of platform-specific
variables and generation of an appropriate set of variables for
compilation (build/build_config.mk) and headers
(build/build_config.h).

port_std is a port module implementation that works on older versions
of gcc/g++ commonly found on Ubuntu, Centos, OSX and Solaris. It is
used with C++0x is not available.

port_osx was removed in favor of port_std, which is more portable.